requirements.txt
Flask==3.0.3
gunicorn==22.0.0
mcp[cli]>=1.2.0
xrpl-py>=4.0.0
pydantic>=2.0.0
psycopg2-binary>=2.9.9
PyJWT==2.9.0
httpx>=0.27.0
google-cloud-secret-manager==2.20.2
